services/api_gateway/requirements.txt
onnxruntime==1.23.2
opencv-python==4.13.0.92
pandas==2.3.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
numpy==2.2.6
seaborn==0.13.2
uvicorn
psycopg2-binary
fastapi
pika
python-multipart
minio